refactor(persistence): move checkpoints + message-content into persistence/ and ratchet the allowlist
