+++ assistant/src/__tests__/plugin-import-boundary-guard.test.ts
@@ -91,11 +91,11 @@ const BASELINE: Record<string, readonly string[]> = {
     "../../../../daemon/conversation-runtime-assembly.js",
     "../../../../daemon/message-types/memory.js",
     "../../../../daemon/trust-context.js",
-    "../../../../memory/checkpoints.js",
+    "../../../../persistence/checkpoints.js",
     "../../../../memory/graph/conversation-graph-memory.js",
     "../../../../memory/memory-marker.js",
     "../../../../memory/memory-recall-log-store.js",
-    "../../../../memory/message-content.js",
+    "../../../../persistence/message-content.js",
     "../../../../memory/prompt-override.js",
     "../../../../memory/v2/cli-command-store.js",
     "../../../../memory/v2/injected-block-slugs.js",
