+++ assistant/src/memory/graph/extraction-job.ts
@@ -7,10 +7,13 @@
 // ---------------------------------------------------------------------------
 
 import type { AssistantConfig } from "../../config/types.js";
+import {
+  getMemoryCheckpoint,
+  setMemoryCheckpoint,
+} from "../../persistence/checkpoints.js";
 import { asString } from "../../persistence/job-utils.js";
 import type { MemoryJob } from "../../persistence/jobs-store.js";
 import { getLogger } from "../../util/logger.js";
-import { getMemoryCheckpoint, setMemoryCheckpoint } from "../checkpoints.js";
 import { runGraphExtraction } from "./extraction.js";
 
 const log = getLogger("graph-extraction-job");
