+++ assistant/src/memory/graph/image-ref-utils.ts
@@ -1,8 +1,8 @@
 import { eq } from "drizzle-orm";
 
 import { getDb } from "../../persistence/db-connection.js";
+import { extractMediaBlocks } from "../../persistence/message-content.js";
 import { messages } from "../../persistence/schema/index.js";
-import { extractMediaBlocks } from "../message-content.js";
 import type { ImageRef } from "./types.js";
 
 /**
