+++ assistant/src/memory/job-handlers/backfill.ts
@@ -1,6 +1,11 @@
 import { and, asc, eq, gt, or } from "drizzle-orm";
 
 import type { AssistantConfig } from "../../config/types.js";
+import {
+  readMessageCursorCheckpoint,
+  resetMessageCursorCheckpoint,
+  writeMessageCursorCheckpoint,
+} from "../../persistence/checkpoints.js";
 import { messageMetadataSchema } from "../../persistence/conversation-crud.js";
 import { getDb } from "../../persistence/db-connection.js";
 import {
@@ -9,11 +14,6 @@ import {
 } from "../../persistence/jobs-store.js";
 import { messages } from "../../persistence/schema/index.js";
 import type { TrustClass } from "../../runtime/actor-trust-resolver.js";
-import {
-  readMessageCursorCheckpoint,
-  resetMessageCursorCheckpoint,
-  writeMessageCursorCheckpoint,
-} from "../checkpoints.js";
 import { indexMessageNow } from "../indexer.js";
 
 const BACKFILL_CHECKPOINT_KEY = "memory:backfill:last_created_at";
