+++ assistant/src/memory/v2/backfill-jobs.ts
@@ -23,10 +23,10 @@ import { getMessages } from "../../persistence/conversation-crud.js";
 import { listConversations } from "../../persistence/conversation-queries.js";
 import { getDb } from "../../persistence/db-connection.js";
 import type { MemoryJob } from "../../persistence/jobs-store.js";
+import { stringifyMessageContent } from "../../persistence/message-content.js";
 import { getLogger } from "../../util/logger.js";
 import { getWorkspaceDir } from "../../util/platform.js";
 import { enqueueEmbedConceptPageJob } from "../jobs/embed-concept-page.js";
-import { stringifyMessageContent } from "../message-content.js";
 import {
   computeOwnActivation,
   selectCandidates,
