+++ assistant/src/persistence/bookmark-crud.ts
@@ -1,8 +1,8 @@
 import { desc, eq } from "drizzle-orm";
 import { v4 as uuid } from "uuid";
 
-import { stringifyMessageContent } from "../memory/message-content.js";
 import type { DrizzleDb } from "./db-connection.js";
+import { stringifyMessageContent } from "./message-content.js";
 import { conversations, messageBookmarks, messages } from "./schema.js";
 
 /**
