+++ assistant/src/persistence/db-maintenance.ts
@@ -1,14 +1,11 @@
 import { statSync } from "node:fs";
 
 import { getConfig } from "../config/loader.js";
-import {
-  getMemoryCheckpoint,
-  setMemoryCheckpoint,
-} from "../memory/checkpoints.js";
 import { getLastUserMessageTimestamp } from "../persistence/conversation-crud.js";
 import { getLogger } from "../util/logger.js";
 import { getDbPath } from "../util/platform.js";
 import { pruneRuns } from "../workflows/journal-store.js";
+import { getMemoryCheckpoint, setMemoryCheckpoint } from "./checkpoints.js";
 import { runAsyncSqlite } from "./db-async-query.js";
 import { getSqlite } from "./db-connection.js";
