+++ assistant/src/persistence/message-content.ts
