+++ assistant/src/plugins/defaults/memory/v3/shadow-plugin.ts
@@ -25,7 +25,6 @@ import { existsSync, readFileSync } from "node:fs";
 import { getConfig } from "../../../../config/loader.js";
 import type { AssistantConfig } from "../../../../config/schema.js";
 import { loadSkillCatalog } from "../../../../config/skills.js";
-import { stringifyMessageContent } from "../../../../memory/message-content.js";
 import { getPageIndex } from "../../../../memory/v2/page-index.js";
 import {
   readPage,
@@ -33,6 +32,7 @@ import {
 } from "../../../../memory/v2/page-store.js";
 import { getMessages } from "../../../../persistence/conversation-crud.js";
 import { getDb, getSqliteFrom } from "../../../../persistence/db-connection.js";
+import { stringifyMessageContent } from "../../../../persistence/message-content.js";
 import { getLogger } from "../../../../util/logger.js";
 import {
   getWorkspaceDir,
