+++ assistant/src/runtime/routes/consolidation-routes.ts
@@ -16,8 +16,8 @@
 import { z } from "zod";
 
 import { getConfig } from "../../config/loader.js";
-import { getMemoryCheckpoint } from "../../memory/checkpoints.js";
 import { MEMORY_V2_CONSOLIDATION_SOURCE } from "../../memory/v2/constants.js";
+import { getMemoryCheckpoint } from "../../persistence/checkpoints.js";
 import {
   getMessageRoleStatsByConversation,
   listConversationsBySource,
