+++ assistant/src/watcher/telemetry.ts
@@ -22,7 +22,7 @@
 import {
   getMemoryCheckpoint,
   setMemoryCheckpoint,
-} from "../memory/checkpoints.js";
+} from "../persistence/checkpoints.js";
 import { recordLifecycleEvent } from "../persistence/lifecycle-events-store.js";
 import { getLogger } from "../util/logger.js";
 import { listWatchers } from "./watcher-store.js";
